+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## Unreleased
 - Fixes an edge case related to binding to events named as `Object.prototype`'s
   properties after having called `off`.
+- Fixes rebinding an inherited event that was previously unbound by calling
+  `off(name)`.
 
 ## 2.0.0 (Feb 13, 2015)
 - You can now inherit from objects and change their event listeners without